refactor: qr_scanner module
