+++ packages/cryptoplease/lib/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart
@@ -10,7 +10,7 @@ import '../../../../core/currency.dart';
 import '../../../../core/presentation/format_amount.dart';
 import '../../../../features/outgoing_direct_payments/presentation/build_context_ext.dart';
 import '../../../../features/outgoing_split_key_payments/bl/bloc.dart';
-import '../../../../features/qr_scanner/qr_scanner_request.dart';
+import '../../../../features/qr_scanner/module.dart';
 import '../../../../l10n/device_locale.dart';
 import '../../../../l10n/l10n.dart';
 import '../../../../ui/dialogs.dart';
