+++ packages/cryptoplease/lib/features/qr_scanner/src/qr_address_data.dart
