+++ packages/cryptoplease/lib/features/qr_scanner/src/qr_scanner_screen.dart
@@ -6,11 +6,11 @@ import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:qr_code_scanner/qr_code_scanner.dart';
 
-import '../../di.dart';
-import '../../gen/assets.gen.dart';
-import '../../l10n/l10n.dart';
-import '../../ui/dialogs.dart';
-import '../../ui/theme.dart';
+import '../../../di.dart';
+import '../../../gen/assets.gen.dart';
+import '../../../l10n/l10n.dart';
+import '../../../ui/dialogs.dart';
+import '../../../ui/theme.dart';
 import 'qr_scanner_bloc.dart';
 
 class QrScannerScreen extends StatelessWidget {
