+++ packages/cryptoplease/lib/features/qr_scanner/src/qr_scanner_state.dart
